# Improve: binary ops with Number Protocol