--- a/vm/src/builtins/dict.rs
+++ b/vm/src/builtins/dict.rs
@@ -479,20 +479,23 @@ impl AsSequence for PyDict {
 
 impl AsNumber for PyDict {
     fn as_number() -> &'static PyNumberMethods {
-        static AS_NUMBER: Lazy<PyNumberMethods> = Lazy::new(|| PyNumberMethods {
-            or: atomic_func!(|num, args, vm| {
-                PyDict::number_downcast(num).or(args.to_pyobject(vm), vm)
+        static AS_NUMBER: PyNumberMethods = PyNumberMethods {
+            or: Some(|num, args, vm| {
+                if let Some(num) = num.obj.downcast_ref::<PyDict>() {
+                    PyDict::or(num, args.to_pyobject(vm), vm)
+                } else {
+                    Ok(vm.ctx.not_implemented())
+                }
             }),
-            inplace_or: atomic_func!(|num, args, vm| {
-                PyDict::ior(
-                    PyDict::number_downcast(num).to_owned(),
-                    args.to_pyobject(vm),
-                    vm,
-                )
-                .map(|d| d.into())
+            inplace_or: Some(|num, args, vm| {
+                if let Some(num) = num.obj.downcast_ref::<PyDict>() {
+                    PyDict::ior(num.to_owned(), args.to_pyobject(vm), vm).map(|d| d.into())
+                } else {
+                    Ok(vm.ctx.not_implemented())
+                }
             }),
             ..PyNumberMethods::NOT_IMPLEMENTED
-        });
+        };
         &AS_NUMBER
     }
 }