--- a/vm/src/builtins/float.rs
+++ b/vm/src/builtins/float.rs
@@ -2,7 +2,6 @@ use super::{
     try_bigint_to_f64, PyByteArray, PyBytes, PyInt, PyIntRef, PyStr, PyStrRef, PyType, PyTypeRef,
 };
 use crate::{
-    atomic_func,
     class::PyClassImpl,
     common::format::FormatSpec,
     common::{float_ops, hash},
@@ -21,7 +20,6 @@ use num_bigint::{BigInt, ToBigInt};
 use num_complex::Complex64;
 use num_rational::Ratio;
 use num_traits::{Signed, ToPrimitive, Zero};
-use once_cell::sync::Lazy;
 
 #[pyclass(module = false, name = "float")]
 #[derive(Debug, Copy, Clone, PartialEq)]
@@ -545,53 +543,32 @@ impl Hashable for PyFloat {
 
 impl AsNumber for PyFloat {
     fn as_number() -> &'static PyNumberMethods {
-        static AS_NUMBER: Lazy<PyNumberMethods> = Lazy::new(|| PyNumberMethods {
-            add: atomic_func!(|num, other, vm| PyFloat::number_op(
-                num,
-                other,
-                |a, b, _vm| a + b,
-                vm
-            )),
-            subtract: atomic_func!(|num, other, vm| PyFloat::number_op(
-                num,
-                other,
-                |a, b, _vm| a - b,
-                vm
-            )),
-            multiply: atomic_func!(|num, other, vm| PyFloat::number_op(
-                num,
-                other,
-                |a, b, _vm| a * b,
-                vm
-            )),
-            remainder: atomic_func!(|num, other, vm| PyFloat::number_op(num, other, inner_mod, vm)),
-            divmod: atomic_func!(|num, other, vm| PyFloat::number_op(num, other, inner_divmod, vm)),
-            power: atomic_func!(|num, other, vm| PyFloat::number_op(num, other, float_pow, vm)),
-            negative: atomic_func!(|num, vm| {
+        static AS_NUMBER: PyNumberMethods = PyNumberMethods {
+            add: Some(|num, other, vm| PyFloat::number_op(num, other, |a, b, _vm| a + b, vm)),
+            subtract: Some(|num, other, vm| PyFloat::number_op(num, other, |a, b, _vm| a - b, vm)),
+            multiply: Some(|num, other, vm| PyFloat::number_op(num, other, |a, b, _vm| a * b, vm)),
+            remainder: Some(|num, other, vm| PyFloat::number_op(num, other, inner_mod, vm)),
+            divmod: Some(|num, other, vm| PyFloat::number_op(num, other, inner_divmod, vm)),
+            power: Some(|num, other, vm| PyFloat::number_op(num, other, float_pow, vm)),
+            negative: Some(|num, vm| {
                 let value = PyFloat::number_downcast(num).value;
                 (-value).to_pyresult(vm)
             }),
-            positive: atomic_func!(
-                |num, vm| PyFloat::number_downcast_exact(num, vm).to_pyresult(vm)
-            ),
-            absolute: atomic_func!(|num, vm| {
+            positive: Some(|num, vm| PyFloat::number_downcast_exact(num, vm).to_pyresult(vm)),
+            absolute: Some(|num, vm| {
                 let value = PyFloat::number_downcast(num).value;
                 value.abs().to_pyresult(vm)
             }),
-            boolean: atomic_func!(|num, _vm| Ok(PyFloat::number_downcast(num).value.is_zero())),
-            int: atomic_func!(|num, vm| {
+            boolean: Some(|num, _vm| Ok(PyFloat::number_downcast(num).value.is_zero())),
+            int: Some(|num, vm| {
                 let value = PyFloat::number_downcast(num).value;
                 try_to_bigint(value, vm).map(|x| vm.ctx.new_int(x))
             }),
-            float: atomic_func!(|num, vm| Ok(PyFloat::number_downcast_exact(num, vm))),
-            floor_divide: atomic_func!(|num, other, vm| {
-                PyFloat::number_op(num, other, inner_floordiv, vm)
-            }),
-            true_divide: atomic_func!(|num, other, vm| {
-                PyFloat::number_op(num, other, inner_div, vm)
-            }),
+            float: Some(|num, vm| Ok(PyFloat::number_downcast_exact(num, vm))),
+            floor_divide: Some(|num, other, vm| PyFloat::number_op(num, other, inner_floordiv, vm)),
+            true_divide: Some(|num, other, vm| PyFloat::number_op(num, other, inner_div, vm)),
             ..PyNumberMethods::NOT_IMPLEMENTED
-        });
+        };
         &AS_NUMBER
     }