--- a/vm/src/protocol/number.rs
+++ b/vm/src/protocol/number.rs
@@ -232,6 +232,7 @@ impl PyNumberMethods {
     }
 }
 
+#[derive(Copy, Clone)]
 pub enum PyNumberBinaryOpSlot {
     Add,
     Subtract,