Closes https://github.com/RustPython/RustPython/issues/4614
Based on https://github.com/RustPython/RustPython/pull/4139
Related to https://github.com/RustPython/RustPython/issues/3244

This PR will not pass some tests at this moment, since Number Protocol is missing for some types.
Let me create separate PRs to implement (part of) missing Number Protocols, and then rebase this PR once those are merged.

closes #4639
closes #4672