@youknowone 2023-03-02
when rustpython crashes on initialization step, debugger is very useful

https://github.com/RustPython/RustPython/wiki/Debugger-with-VSCode