@youknowone 2023-03-02
The next step is update_slot in slot.rs.
When operator is overriden from Python side, we have to set a function to the python function from rust side.