@youknowone 2023-03-02
To support simple binary operators, looking for `index_wrapper` or `len_wrapper` will be useful. One wrapper matches to one slot.
To support self-mutating unary ops like `__iadd__` and `__add__`, looking for `richcompare_wrapper` will be useful. When user defined `__add__` on python side, both `__add__` and `__iadd__` slots have to be set.(1 python function to n slots) `richcompare_wrapper` is doing similar work - but the opposite way. (n python funciton to 1 slot)