@xiaozhiyan 2023-03-05
The following is the behavior in CPython
```
>>> class C:
...     def __add__(self, x):
...             return x + 100
...     def __radd__(self, x):
...             return x + 200
... 
>>> c = C()
>>> c + 5
105
>>> 5 + c
205
```