@youknowone 2023-03-06
I finally understand the trick. instead of `number_binary_op_wrapper`, adding `number_binary_op_left_wrapper` and `number_binary_op_right_wrapper` allows to share single PyNumber operation. And they can be separately overriden from python side. We need another `PyNumberSlots` struct.