@youknowone 2023-03-06
@xiaozhiyan I made a rough sketch of the cpython way
https://github.com/RustPython/RustPython/commit/10c1929eb00953b04e53f33c7ffac600f0487225