@youknowone 2023-03-06
the difference between my version and cpython is slot implementation.
I used different slot function for wrapper and number protocol implementation. But cpython use a single function for both cases
https://github.com/python/cpython/blob/6716254e71eeb4666fd6d1a13857832caad7b19f/Objects/typeobject.c#L7792-L7833