@youknowone 2023-03-07
Because the patch is really huge, I’d like to suggest to split it down to multiple partial fixes. 

Currently `vm.binary_{i}op{1}` with `#[allow(dead_code)]` seems to be a chunk.
Once finished, `update_slot!` stuff can be a chunk too. Even if it is not very correct at the moment, it will be fine unless we actually call them from `vm._add` etc

@qingshi163 how do you think?