@qingshi163 2023-03-07
I am bit confuse about the right_ops witch did not exist in cpython, why can't we just swap the direction as what we doing before?