@xiaozhiyan 2023-03-07
> I am bit confuse about the right_ops witch did not exist in cpython, why can't we just swap the direction as what we doing before?

As in the example https://github.com/RustPython/RustPython/pull/4615#issuecomment-1455131249 , left and binary ops and be defined separately. So we need to store them separately. Please check the latest commit https://github.com/RustPython/RustPython/pull/4615/commits/cdac5cfb69ee5bc976164b1da05d2b0ebac2ca1e to create `PyNumberSlots` to save both left and right binary ops.