@xiaozhiyan 2023-03-07
I temporarily implemented Number Protocol for `PyBool` in commit https://github.com/RustPython/RustPython/pull/4615/commits/22fca357c587ff9c0315ff11b14c03c6e1ab7885  to see whether this PR can pass all the tests.

- As discussed in https://github.com/RustPython/RustPython/pull/4639 , Number Protocol of `PyBool` will be implemented by inheriting slots of `PyInt`, which depends on this PR. 
- This PR also requires Number Protocol of `PyBool` to pass tests.

So let me deprecate the PR https://github.com/RustPython/RustPython/pull/4639 and implement it within this PR.