@xiaozhiyan 2023-03-07
As in https://github.com/RustPython/RustPython/actions/runs/4357591970/jobs/7617136283 , 
```
2 tests failed:
    test_array test_bytes
```

```
TypeError: '*' not supported between instances of 'int' and 'array'
```

```
RUST_BACKTRACE=1 cargo run Lib/test/test_bytes.py

thread 'main' panicked at 'assertion failed: self.payload_is::<T>()', vm/src/object/core.rs:730:9
```