@youknowone 2023-03-07
`mod` operator of `bytes` and `array` is not communicative like numeric types. their mod and rmod are different.