@youknowone 2023-03-08
I found why array doesn't work. It missed update_slot. I can fix it tomrorow.