@xiaozhiyan 2023-03-08
Since related changes in this PR seem to increase along the way, now it may be a good timing to move the following loosely coupled part to another PR as prerequisite of the current one.

- Refactor all existing Number Protocol methods (actually only some of them need big changes)
  - so that they accept arguments of type similar to (&PyObject, &PyObject, &VirtualMachine)
  - within this method, we don't need to consider whether the caller comes from left or right binary op, and the method just needs to handle two `PyObject` arguments and return the value (or not_implemented) as expected

- Then these methods can be used by both left and right binary ops in the current PR

@youknowone If you also agree with the approach, I will open a new PR to work on that as prerequisite of the current PR.