@xiaozhiyan 2023-03-09
From latest CI after the temp fixes,

- there's a minor clippy warning, and I've fixed it locally

- the following tests failed
```
test_collections test_dict test_dictviews test_types
```

- `test_collections`
  - methods defined in `ChainMap` like `__or__` not taking effect

- `test_dict`
  - need to implement Number Protocol for `PyDictItems`
```
'^' not supported between instances of 'dict_items' and 'dict_items'
```

- `test_dictviews`
  - need to implement Number Protocol for `PyDictKeys`
```
'&' not supported between instances of 'dict_keys' and 'list_iterator'
```

- `test_types`
```
'|' not supported between instances of 'GenericAlias' and 'GenericAlias'
```