@youknowone 2023-03-09
I agree we need to remove PyNumber from PyNumberMethods.
But it will last on PyNumberSlots.

I will take a look how we can handle it smart way.

@qingshi163 because this PR is already huge, I'd like to merge it unless it has obvious wrong design or bugs. could you review it?