@xiaozhiyan 2023-03-09
@youknowone 
The latest CI failed with Clippy error, but that part of source has not been changed for a long time. Why suddenly CI treats it as an error and makes the CI fail?
https://github.com/RustPython/RustPython/actions/runs/4375751891/jobs/7656913583