@xiaozhiyan 2023-03-13
Measure performance enhancement of this PR

- Download https://github.com/python/pyperformance/blob/main/pyperformance/data-files/benchmarks/bm_nbody/run_benchmark.py to `benches/benchmarks/bm_nbody.py`

- Before patch
```
% git log -n1
commit 1fceeab0fcc77c11dfe94092b27b2164b94d62e4 (HEAD)
Merge: 51020cd63 4d05077ec
Author: Jeong YunWon <69878+youknowone@users.noreply.github.com>
Date:   Thu Mar 9 21:19:47 2023 +0900

    Merge pull request #4674 from carlosmiei/copy-update
    
    Update copy.py from CPython 3.11

% cargo run --release benches/benchmarks/bm_nbody.py -o base.json
.....................
nbody: Mean +- std dev: 4.01 sec +- 0.17 sec

% cargo run --release benches/benchmarks/bm_nbody.py -o base2.json
.....................
nbody: Mean +- std dev: 3.91 sec +- 0.14 sec
```

- After patch
```
% git log -n1
commit ea665cb743ef1f2a1a2b6030e8384d7ab814b305 (HEAD)
Merge: 1fceeab0f 228772020
Author: Jeong YunWon <69878+youknowone@users.noreply.github.com>
Date:   Fri Mar 10 02:47:01 2023 +0900

    Merge pull request #4615 from xiaozhiyan/binaryops-with-number-protocol
    
    Improve: binary ops with Number Protocol

% cargo run --release benches/benchmarks/bm_nbody.py -o patch.json
.....................
nbody: Mean +- std dev: 1.24 sec +- 0.08 sec

% cargo run --release benches/benchmarks/bm_nbody.py -o patch2.json
.....................
nbody: Mean +- std dev: 1.28 sec +- 0.12 sec
```

- CPython 3.11
```
% python -V
Python 3.11.0

% python benches/benchmarks/bm_nbody.py -o cpython.json
.....................
nbody: Mean +- std dev: 101 ms +- 2 ms

% python benches/benchmarks/bm_nbody.py -o cpython2.json
.....................
nbody: Mean +- std dev: 104 ms +- 7 ms
```

- Performance comparison
```
% pyperf compare_to cpython.json patch.json base.json --table 
+-----------+---------+-------------------------+-------------------------+
| Benchmark | cpython | patch                   | base                    |
+===========+=========+=========================+=========================+
| nbody     | 101 ms  | 1.24 sec: 12.28x slower | 4.01 sec: 39.79x slower |
+-----------+---------+-------------------------+-------------------------+

% pyperf compare_to cpython2.json patch2.json base2.json --table
+-----------+----------+-------------------------+-------------------------+
| Benchmark | cpython2 | patch2                  | base2                   |
+===========+==========+=========================+=========================+
| nbody     | 104 ms   | 1.28 sec: 12.34x slower | 3.91 sec: 37.54x slower |
+-----------+----------+-------------------------+-------------------------+
```